requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn==0.29.0

# Database
sqlalchemy==2.0.30
aiosqlite==0.20.0

# AI / LLM
langchain==0.2.1
langchain-community==0.2.1
langgraph==0.1.1
ollama==0.2.0

# Sentiment Analysis – RoBERTa (cardiffnlp/twitter-roberta-base-sentiment-latest)
# Handles conversational/complaint language far better than FinBERT for customer chat.
# torch and transformers are already installed via openai-whisper dependencies.
transformers>=4.30.0
torch>=2.0.0
rapidfuzz>=3.0.0

# Speech-to-Text (Voice Call Analysis)
# NOTE: openai-whisper==20231117 has a broken setup.py on newer pip.
# Install manually with: pip install openai-whisper
# (already installed if you see it in pip list)
ffmpeg-python==0.2.0

# Sarvam AI SDK (Official - for Speech-to-Text & Translation)
sarvamai>=0.1.0

# Vector DB
chromadb==0.5.0

# Embeddings
sentence-transformers==3.0.0

# Data & Analytics
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
python-multipart==0.0.9

# Auth
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
